Update GrovePi with database insert

+++ Projects/Sensors_GrovePi/mysql/Create.sql
@@ -1,9 +1,9 @@
 CREATE TABLE Zaznam(
     `Id` int (11) NOT NULL AUTO_INCREMENT,
     `Time` datetime NOT NULL,
-    `Temp` float(6,3),
-    `Hum` int(6),
-    `Lux` int(6),
-    `Noise` int(6),
+    `Temp` float(20,6),
+    `Hum` float(20,6),
+    `Lux` float(20,6),
+    `Noise` float(20,6),
     PRIMARY KEY (`Id`)
 );
